requirements.txt
pillow==10.4.0
python-stdnum==1.20
sqlparse==0.5.1
tzdata==2024.1
